Model: Only store axioms in axioms_, make properties/methods etc just views of axioms_.
This contains a modified bootstrap mechanism which is more axiom
based.  The early installModel support does not contain knowledge of
properties/methods pseudo properties.

This makes it more viable to edit models in a GUI or programmatically
as updates to model properties will not blindly append to axioms_ evey
postSet.