Re-work slotE_() slightly.
Fire the listener on load if it isn't the first load.  Since we
unsubscribe when unloading we need to recompute nextE() next time we
load since we may have missed events to update it.

Remove debugger statment from top of listener.  It is possible for a
listener to fire even when not in the LOADED state because the
listener is merged.  In such a situation we just ignore the event,
because we will fire it again when we're actually loaded.